Split player.js into player, enemy and entity
